refactor: rename theme helpers to color-scheme
