+++ app/helpers/ui/color-scheme-init.entry.ts
@@ -1,0 +1,3 @@
+import { initColorScheme } from "./color-scheme"
+
+initColorScheme()
