--- app/helpers/ui/theme-init.entry.ts
@@ -1,3 +1,0 @@
-import { initColorScheme } from "./theme"
-
-initColorScheme()
