+++ app/routes/_renderer.tsx
@@ -2,7 +2,7 @@ import { Style } from "hono/css"
 import { jsxRenderer, useRequestContext } from "hono/jsx-renderer"
 import { Link, Script } from "honox/server"
 import Toast from "../components/ui/toast"
-import themeInitScript from "../helpers/ui/theme-init.entry.ts?js"
+import colorSchemeInitScript from "../helpers/ui/color-scheme-init.entry.ts?js"
 import { deleteToastCookie, getToastCookie } from "../helpers/ui/toast"
 
 export default jsxRenderer(({ children }) => {
@@ -18,8 +18,8 @@ export default jsxRenderer(({ children }) => {
       <head>
         <meta charset="utf-8" />
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-        {/* biome-ignore lint/security/noDangerouslySetInnerHtml: must inline pre-render theme script */}
-        <script dangerouslySetInnerHTML={{ __html: themeInitScript }} />
+        {/* biome-ignore lint/security/noDangerouslySetInnerHtml: must inline pre-render color scheme script */}
+        <script dangerouslySetInnerHTML={{ __html: colorSchemeInitScript }} />
         <link rel="icon" href="/favicon.ico" />
         <Link href="/app/style.css" rel="stylesheet" />
         <Script src="/app/client.ts" async />
